requirements.txt
bcrypt==4.0.1
email_validator==2.2.0
fastapi==0.116.0
passlib==1.7.4
pydantic==2.11.7
python-dotenv==1.1.1
python-jose==3.5.0
SQLAlchemy==2.0.41
uvicorn==0.35.0